+++ src/hardis-status-provider.ts
@@ -17,7 +17,7 @@ export class HardisStatusProvider
   getChildren(element?: StatusTreeItem): Thenable<StatusTreeItem[]> {
     if (!this.workspaceRoot) {
       vscode.window.showInformationMessage(
-        "No info available until you open a Salesforce project"
+        "🦙 No info available until you open a Salesforce project"
       );
       return Promise.resolve([]);
     }
@@ -142,15 +142,15 @@ export class HardisStatusProvider
         if (daysBeforeExpiration < 0) {
           item.icon = "warning-red.svg";
           item.tooltip = `You org expired on ${orgInfo.expirationDate}. You need to create a new one.`;
-          vscode.window.showErrorMessage(item.tooltip);
+          vscode.window.showErrorMessage(`🦙 ${item.tooltip}`,"Close");
         } else if (daysBeforeExpiration < 3) {
           item.icon = "warning-red.svg";
           item.tooltip = `You scratch org will expire in ${daysBeforeExpiration} days !!! Save your scratch org content and create a new one or your work will be lost !!!`;
-          vscode.window.showErrorMessage(item.tooltip);
+          vscode.window.showErrorMessage(`🦙 ${item.tooltip}`,"Close");
         } else if (daysBeforeExpiration < 7) {
           item.icon = "warning.svg";
           item.tooltip = `Your scratch org will expire in ${daysBeforeExpiration} days. You should soon create a new scratch org to avoid loosing your work`;
-          vscode.window.showWarningMessage(item.tooltip);
+          vscode.window.showWarningMessage(`🦙 ${item.tooltip}`,"Close");
         }
         items.push(item);
       }